requirements.txt
customtkinter>=5.2.0
sqlalchemy>=2.0.0
python-barcode>=0.15.1
Pillow>=10.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
requests>=2.32.0
Jinja2>=3.1.6
schedule>=1.2.0
qrcode[pil]>=7.4.2
Flask>=3.0.0
Flask-CORS>=4.0.0
# ASGI server للاستضافة
uvicorn[standard]>=0.24.0
asgiref>=3.7.0
# دعم قواعد البيانات
psycopg2-binary>=2.9.0  # PostgreSQL (للخوادم السحابية)
pymysql>=1.1.0  # MySQL (اختياري)
# مكتبات النسخ الاحتياطي إلى Google Drive (اختيارية)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
# لبناء Windows EXE
pyinstaller>=6.0.0
